requirements.txt
pandas
word2number
flask
fastapi
uvicorn
requests
python-dotenv
beautifulsoup4
lxml
